+++ CHANGELOG.md
@@ -11,12 +11,15 @@
 - `TabBuilder` for the `BuiltTab`
 - Support for all implementations of `Into<WidgetText>` in tab titles
 - Style editor in the `hello` example
+- Added `Tree::find_tab`, `TabViewer`, `DynamicTabViewer`, `DynamicTree`
+- Added a `text_editor` example
 
 ## Changed
 
 - If a tab is dropped onto the tab bar it will be inserted into the index that it is dropped onto.
 - Now when you drag a tab it has an outline along the entire length of the edges of it
 - Bumped MSRV to `1.62`
+- `Tree` is now generic over how you want to represent a tab
 
 ## Breaking changes
 
@@ -25,6 +28,7 @@
 - Renamed `Style::separator_size` to `Style::separator_width`
 - Removed `Style::tab_text_color` as you can now set the tab text color of a tab by passing `RichText` for its title
 - Removed the requirement of creating your own Context type
+- Renamed `Tree::set_focused` to `Tree::set_focused_node`
 
 ## Fixed